server/requirements.txt
Django>=5.2.2,<5.3.0
django-cors-headers>=4.3.1,<4.4.0
django-filter>=23.5,<24.0
djangorestframework>=3.14.0,<3.15.0
djangorestframework-simplejwt>=5.3.1,<5.4.0
Pillow>=10.1.0,<10.2.0
pytz>=2023.3
sqlparse>=0.4.4,<0.5.0